+++ src/idunno.Bluesky/Graph/Follow.cs
@@ -5,14 +5,16 @@ using System.Text.Json.Serialization;
 
 using idunno.AtProto;
 using idunno.AtProto.Repo;
+using idunno.Bluesky.Record;
 
-namespace idunno.Bluesky.Record;
+namespace idunno.Bluesky.Graph;
 
 /// <summary>
 /// Encapsulates a follow record.
 /// </summary>
-[JsonPolymorphic(IgnoreUnrecognizedTypeDiscriminators = true,
-                 UnknownDerivedTypeHandling = JsonUnknownDerivedTypeHandling.FallBackToNearestAncestor)]
+[JsonPolymorphic(
+    IgnoreUnrecognizedTypeDiscriminators = true,
+    UnknownDerivedTypeHandling = JsonUnknownDerivedTypeHandling.FallBackToNearestAncestor)]
 [JsonDerivedType(typeof(Follow), RecordType.Follow)]
 public record Follow : BlueskyTimestampedRecord
 {
@@ -20,9 +22,20 @@ public record Follow : BlueskyTimestampedRecord
     /// Creates a new instance of <see cref="Follow"/> with <see cref="BlueskyTimestampedRecord.CreatedAt"/> set to the current date and time.
     /// </summary>
     /// <param name="subject">The <see cref="Did"/> to the actor to be followed.</param>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="subject"/> is <see langword="null"/>.</exception>
+    public Follow(Did subject) : this(subject: subject, via: null, createdAt: DateTimeOffset.UtcNow)
+    {
+        ArgumentNullException.ThrowIfNull(subject);
+    }
+
+
+    /// <summary>
+    /// Creates a new instance of <see cref="Follow"/> with <see cref="BlueskyTimestampedRecord.CreatedAt"/> set to the current date and time.
+    /// </summary>
+    /// <param name="subject">The <see cref="Did"/> to the actor to be followed.</param>
     /// <param name="via">A <see cref="StrongReference"/> to the starter pack the follow originated from, if any.</param>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="subject"/> is <see langword="null"/>.</exception>
-    public Follow(Did subject, StrongReference? via = null) : this(subject: subject, via: via, createdAt: DateTimeOffset.UtcNow)
+    public Follow(Did subject, StrongReference? via) : this(subject: subject, via: via, createdAt: DateTimeOffset.UtcNow)
     {
         ArgumentNullException.ThrowIfNull(subject);
     }
@@ -57,6 +70,7 @@ public record Follow : BlueskyTimestampedRecord
     /// Gets the <see cref="Did"/> of the actor being followed.
     /// </summary>
     [JsonInclude]
+    [JsonRequired]
     public Did Subject { get; init; }
 
     /// <summary>