+++ src/idunno.Bluesky/Record/BlueskyAgent.cs
@@ -414,9 +414,9 @@ public partial class BlueskyAgent
     }
 
     /// <summary>
-    /// Creates a <see cref="BlueskyList"/>.
+    /// Creates a <see cref="List"/>.
     /// </summary>
-    /// <param name="list">The <see cref="BlueskyList"/> to create.</param>
+    /// <param name="list">The <see cref="List"/> to create.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns>The task object representing the asynchronous operation.</returns>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="list"/> is <see langword="null"/>.</exception>
@@ -429,7 +429,7 @@ public partial class BlueskyAgent
         "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
         Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
     public async Task<AtProtoHttpResult<CreateRecordResult>> CreateList(
-        BlueskyList list,
+        List list,
         CancellationToken cancellationToken = default)
     {
         ArgumentNullException.ThrowIfNull(list);
@@ -474,77 +474,6 @@ public partial class BlueskyAgent
     }
 
     /// <summary>
-    /// Updates the referenced list record.
-    /// </summary>
-    /// <param name="list">The <see cref="AtProtoRepositoryRecord{TRecord}"/> referenced <see cref="BlueskyList"/> to update.</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="list"/> or its Uri, or the URI Collection or RecordKey property is <see langword="null"/>.</exception>
-    /// <exception cref="ArgumentException">Thrown if <paramref name="list"/> does not point to a list.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    public async Task<AtProtoHttpResult<PutRecordResult>> UpdateList(
-        AtProtoRepositoryRecord<BlueskyList> list,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(list);
-        ArgumentNullException.ThrowIfNull(list.Uri);
-        ArgumentNullException.ThrowIfNull(list.Uri.Collection);
-        ArgumentNullException.ThrowIfNull(list.Uri.RecordKey);
-        ArgumentOutOfRangeException.ThrowIfNotEqual(list.Uri.Collection, CollectionNsid.List);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await UpdateList(
-            uri: list.Uri,
-            list: list.Value,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Updates the list record referenced by its <paramref name="uri"/>.
-    /// </summary>
-    /// <param name="uri">The <see cref="AtUri"/> of the list record to update.</param>
-    /// <param name="list">The <see cref="BlueskyList"/> to update the record with</param>
-    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
-    /// <returns>The task object representing the asynchronous operation.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/> or its Collection or RecordKey property is <see langword="null"/>.</exception>
-    /// <exception cref="ArgumentException">Thrown if <paramref name="uri"/> does not point to a list.</exception>
-    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
-    [UnconditionalSuppressMessage(
-        "Trimming",
-        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    [UnconditionalSuppressMessage("AOT",
-        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
-        Justification = "All types are preserved in the JsonSerializerOptions call to Put().")]
-    public async Task<AtProtoHttpResult<PutRecordResult>> UpdateList(
-        AtUri uri,
-        BlueskyList list,
-        CancellationToken cancellationToken = default)
-    {
-        ArgumentNullException.ThrowIfNull(uri);
-        ArgumentNullException.ThrowIfNull(uri.Collection);
-        ArgumentNullException.ThrowIfNull(uri.RecordKey);
-        ArgumentOutOfRangeException.ThrowIfNotEqual(uri.Collection, CollectionNsid.List);
-
-        if (!IsAuthenticated)
-        {
-            throw new AuthenticationRequiredException();
-        }
-
-        return await PutRecord<BlueskyTimestampedRecord>(
-            record: list,
-            collection: CollectionNsid.List,
-            rKey: uri.RecordKey,
-            jsonSerializerOptions: BlueskyServer.BlueskyJsonSerializerOptions,
-            validate: true,
-            cancellationToken: cancellationToken).ConfigureAwait(false);
-    }
-
-    /// <summary>
     /// Adds the <paramref name="did"/> to the specified <paramref name="uri"/>.
     /// </summary>
     /// <param name="uri">The <see cref="AtUri"/> of the list to add the <paramref name="did"/> to.</param>
@@ -573,7 +502,7 @@ public partial class BlueskyAgent
             throw new AuthenticationRequiredException();
         }
 
-        BlueskyListItem listItem = new() { List = uri, Subject = did };
+        ListItem listItem = new() { List = uri, Subject = did };
 
         return await CreateRecord<BlueskyTimestampedRecord>(
             record: listItem,