+++ src/idunno.Bluesky/Record/BlueskyRecord.cs
@@ -4,6 +4,7 @@
 using System.Text.Json.Serialization;
 
 using idunno.AtProto.Repo;
+using idunno.Bluesky.Graph;
 
 namespace idunno.Bluesky.Record;
 
@@ -21,10 +22,12 @@ namespace idunno.Bluesky.Record;
 [JsonDerivedType(typeof(StarterPack), RecordType.StarterPack)]
 [JsonDerivedType(typeof(LabelerDeclaration), RecordType.LabelerDeclaration)]
 [JsonDerivedType(typeof(Verification), RecordType.Verification)]
-[JsonDerivedType(typeof(BlueskyList), RecordType.List)]
-[JsonDerivedType(typeof(BlueskyListItem), RecordType.ListItem)]
+[JsonDerivedType(typeof(List), RecordType.List)]
+[JsonDerivedType(typeof(ListItem), RecordType.ListItem)]
 [JsonDerivedType(typeof(Notifications.Declaration), RecordType.NotificationDeclaration)]
 [JsonDerivedType(typeof(Status), RecordType.Status)]
+[JsonDerivedType(typeof(ReferenceListOptOut), RecordType.ReferenceListOptOut)]
+[JsonDerivedType(typeof(ListBlock), RecordType.ListBlock)]
 public record BlueskyRecord : AtProtoRecord
 {
     /// <summary>